upbai

+++ src/ss12_java_collections_framework/bai_tap/luyen_tap/SapXepGiamDanTheoGia.java
@@ -6,6 +6,6 @@ public class SapXepGiamDanTheoGia implements Comparator<Product> {
 
     @Override
     public int compare(Product o1, Product o2) {
-        return o2.getPrice()-o1.getPrice();
+        return (int) (o2.getPrice()-o1.getPrice());
     }
 }
