+++ src/ss12_java_collections_framework/bai_tap/luyen_tap/SapXepTangDanTheoGia.java
@@ -6,6 +6,6 @@ public class SapXepTangDanTheoGia implements Comparator<Product> {
 
     @Override
     public int compare(Product o1, Product o2) {
-        return o1.getPrice()-o2.getPrice();
+        return (int) (o1.getPrice()-o2.getPrice());
     }
 }
